I am such a splendid developer
I fixed so much with this commit it's like magic. The app drawer is flawless AND works flawlessly. It's even smart enough that a press of the back button will get ya out of the app when you want to but will kick you right back to your homescreen first when you are just exploring different fragments. So wow amirite??

Also a couple of small bugfixes, some Jan-code cleaned up and some styles beautifulized.

Peace