This is a fabulous commit
Say hello to the icons for the beta and complete version of the app!

Also, by implementing a final check for network connectivity, the app is now smart enough to know wheter it has to work with local data or can go online -- or can't do shit because there is no local data.

Peace